Settings for zappa deployment
