+++ README.MD
@@ -6,12 +6,25 @@ for generating timereport deviations
 Create a .env file with TOKEN information for your slack app
 
 ```
+# slack token variables
 SLACK_BOT_TOKEN="xxxxxxx"
 SLACK_VERIFICATION_TOKEN="xxxxxx"
+# flask listen port
 LISTEN_PORT="5000"
+# mongodb environment variables
+MONGO_HOST="hostname.example.com"
+MONGO_PORT="27017"
+MONGO_USERNAME="dbuser"
+MONGO_PASSWORD="dbpassword"
+MONGO_DBNAME="dbname"
 ```
 
 ```
-python main.py
-```
+# information
 
+this particular setup runs on AWS lambda using api Gateway.
+
+Deployed through zappa.
+
+pip install zappa
+```
