+++ zappa_settings.json
@@ -1,0 +1,26 @@
+{
+    "timereport": {
+        "app_function": "timereport.app", 
+        "profile_name": "default", 
+        "project_name": "slack_timereport", 
+        "runtime": "python3.6", 
+        "s3_bucket": "timereport.codelabs.se",
+        "aws_region": "us-east-1"
+    }
+    "timereport_fetch": {
+        "app_function": "timereport_fetch.app", 
+        "profile_name": "default", 
+        "project_name": "slack_timereport_fetch", 
+        "runtime": "python3.6", 
+        "s3_bucket": "timereport.codelabs.se",
+        "aws_region": "us-east-1"
+    }
+    "message_actions": {
+        "app_function": "message_actions.app", 
+        "profile_name": "default", 
+        "project_name": "message_actions", 
+        "runtime": "python3.6", 
+        "s3_bucket": "timereport.codelabs.se",
+        "aws_region": "us-east-1"
+    }
+}
